backend/requirements.txt
Django==5.2.8
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
psycopg2-binary
Pillow
django-allauth
PyJWT